+++ exporters/autoexport/logs_test.go
@@ -9,6 +9,8 @@ import (
 	"reflect"
 	"testing"
 
+	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp"
+
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/otel/exporters/stdout/stdoutlog"
@@ -17,8 +19,9 @@ import (
 
 func TestLogExporterNone(t *testing.T) {
 	t.Setenv("OTEL_LOGS_EXPORTER", "none")
-	got, err := NewLogExporter(context.Background())
+	exporters, err := NewLogExporters(context.Background())
 	assert.NoError(t, err)
+	got := exporters[0]
 	t.Cleanup(func() {
 		assert.NoError(t, got.ForceFlush(context.Background()))
 		assert.NoError(t, got.Shutdown(context.Background()))
@@ -29,8 +32,10 @@ func TestLogExporterNone(t *testing.T) {
 
 func TestLogExporterConsole(t *testing.T) {
 	t.Setenv("OTEL_LOGS_EXPORTER", "console")
-	got, err := NewLogExporter(context.Background())
+	exporters, err := NewLogExporters(context.Background())
 	assert.NoError(t, err)
+
+	got := exporters[0]
 	assert.IsType(t, &stdoutlog.Exporter{}, got)
 }
 
@@ -46,8 +51,9 @@ func TestLogExporterOTLP(t *testing.T) {
 		t.Run(fmt.Sprintf("protocol=%q", tc.protocol), func(t *testing.T) {
 			t.Setenv("OTEL_EXPORTER_OTLP_PROTOCOL", tc.protocol)
 
-			got, err := NewLogExporter(context.Background())
+			exporters, err := NewLogExporters(context.Background())
 			assert.NoError(t, err)
+			got := exporters[0]
 			t.Cleanup(func() {
 				assert.NoError(t, got.Shutdown(context.Background()))
 			})
@@ -72,7 +78,8 @@ func TestLogExporterOTLPWithDedicatedProtocol(t *testing.T) {
 		t.Run(fmt.Sprintf("protocol=%q", tc.protocol), func(t *testing.T) {
 			t.Setenv("OTEL_EXPORTER_OTLP_LOGS_PROTOCOL", tc.protocol)
 
-			got, err := NewLogExporter(context.Background())
+			exporters, err := NewLogExporters(context.Background())
+			got := exporters[0]
 			assert.NoError(t, err)
 			t.Cleanup(func() {
 				assert.NoError(t, got.Shutdown(context.Background()))
@@ -86,10 +93,46 @@ func TestLogExporterOTLPWithDedicatedProtocol(t *testing.T) {
 	}
 }
 
+func TestLogExporterOTLPMultiple(t *testing.T) {
+	t.Setenv("OTEL_LOGS_EXPORTER", "otlp,console")
+	t.Setenv("OTEL_EXPORTER_OTLP_PROTOCOL", "http/protobuf")
+
+	exporters, err := NewLogExporters(context.Background())
+	assert.NoError(t, err)
+	assert.Len(t, exporters, 2)
+
+	assert.Implements(t, new(log.Exporter), exporters[0])
+	assert.IsType(t, &otlploghttp.Exporter{}, exporters[0])
+
+	assert.Implements(t, new(log.Exporter), exporters[1])
+	assert.IsType(t, &stdoutlog.Exporter{}, exporters[1])
+
+	t.Cleanup(func() {
+		assert.NoError(t, exporters[0].Shutdown(context.Background()))
+		assert.NoError(t, exporters[1].Shutdown(context.Background()))
+	})
+}
+
+func TestLogExporterOTLPMultiple_FailsIfOneValueIsInvalid(t *testing.T) {
+	t.Setenv("OTEL_LOGS_EXPORTER", "otlp,something")
+	t.Setenv("OTEL_EXPORTER_OTLP_PROTOCOL", "http/protobuf")
+
+	_, err := NewLogExporters(context.Background())
+	assert.Error(t, err)
+}
+
 func TestLogExporterOTLPOverInvalidProtocol(t *testing.T) {
 	t.Setenv("OTEL_LOGS_EXPORTER", "otlp")
 	t.Setenv("OTEL_EXPORTER_OTLP_PROTOCOL", "invalid-protocol")
 
-	_, err := NewLogExporter(context.Background())
+	_, err := NewLogExporters(context.Background())
 	assert.Error(t, err)
+}
+
+func TestLogExporterDeprecatedNewLogExporterReturnsTheFirstExporter(t *testing.T) {
+	t.Setenv("OTEL_LOGS_EXPORTER", "console,otlp")
+	got, err := NewLogExporter(context.Background())
+
+	assert.NoError(t, err)
+	assert.IsType(t, &stdoutlog.Exporter{}, got)
 }