+++ exporters/autoexport/signal_test.go
@@ -16,14 +16,14 @@ func TestOTLPExporterReturnedWhenNoEnvOrFallbackExporterConfigured(t *testing.T)
 	assert.NoError(t, ts.registry.store("otlp", factory("test-otlp-exporter")))
 	exp, err := ts.create(context.Background())
 	assert.NoError(t, err)
-	assert.Equal(t, exp.string, "test-otlp-exporter")
+	assert.Equal(t, exp[0].string, "test-otlp-exporter")
 }
 
 func TestFallbackExporterReturnedWhenNoEnvExporterConfigured(t *testing.T) {
 	ts := newSignal[*testType]("TEST_TYPE_KEY")
 	exp, err := ts.create(context.Background(), withFallbackFactory(factory("test-fallback-exporter")))
 	assert.NoError(t, err)
-	assert.Equal(t, exp.string, "test-fallback-exporter")
+	assert.Equal(t, exp[0].string, "test-fallback-exporter")
 }
 
 func TestFallbackExporterFactoryErrorReturnedWhenNoEnvExporterConfiguredAndFallbackFactoryReturnsAnError(t *testing.T) {
@@ -48,5 +48,5 @@ func TestEnvExporterIsPreferredOverFallbackExporter(t *testing.T) {
 
 	exp, err := ts.create(context.Background(), withFallbackFactory(factory("test-fallback-exporter")))
 	assert.NoError(t, err)
-	assert.Equal(t, exp.string, "test-env-exporter")
+	assert.Equal(t, exp[0].string, "test-env-exporter")
 }